pyproject.toml
[tool.poetry]
name = "soen341"
version = "0.1.0"
description = ""
authors = ["MinhTran <minhtrannhat2001@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.4.1"
cfgv = "3.3.1"
distlib = "0.3.4"
Django = "4.0.1"
filelock = "3.4.2"
identify = "2.4.4"
nodeenv = "1.6.0"
platformdirs = "2.4.1"
pre-commit = "2.17.0"
PyYAML = "6.0"
six = "1.16.0"
sqlparse = "0.4.2"
toml = "0.10.2"
virtualenv = "20.13.0"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.11.0"
djangorestframework-simplejwt = "^5.1.0"
django-allauth = "^0.50.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
